+++ src/finlang.gen/GenSimNumerics.cs
@@ -1,7 +1,5 @@
-using System;
-using System.Reflection;
+using finlang.Utils.StringUtilExtensions;
 using Xunit;
-using System.Linq;
 
 namespace finlang.gen;