+++ src/finlang.test/TranspilerTest/InterfaceTests.cs
@@ -1,5 +1,5 @@
 ﻿using finlang.test.Output;
-using finlang.Transpiler;
+using finlang.Utils;
 
 namespace finlang.test.TranspilerTest;