+++ src/finlang/Transpiler/CFileGenerator.cs
@@ -1,3 +1,4 @@
+using finlang.Utils;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;