+++ src/finlang/Transpiler/CTranspiler.cs
@@ -1,4 +1,5 @@
 ﻿using finlang.Output;
+using finlang.Utils;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;