+++ CHANGELOG.md
@@ -22,6 +22,10 @@ You can easily find changes by searching for `"### Changed"` in the in this mark
 # Releases
 Test/interim releases are not documented here.
 
+## [wip]
+### Added
+- Add StringUtils indent functions
+
 ---
 ## [0.5.0]
 ### Changed (BREAKING-CHANGES)